+++ State/StateCommunicator.py
@@ -3,7 +3,6 @@ from Server.WebsocketClientHandlerRegistry import GAMES
 from QueueEntries.Sendable import Sendable, ErrorSendable
 from State.States import *
 from State.StateCommunicatorInterface import StateCommunicatorInterface
-import inspect
 
 # type hints
 from typing import Dict
